Fix for copy methods on huge streams (>2 mio. bytes/chars). Added lacking copy methods for char streams.

+++ src/de/diddiz/utils/ProgressListener.java
@@ -5,5 +5,5 @@ public interface ProgressListener
 	/**
 	 * @param position Current position in the task. Upper bound is not necessarily known.
 	 */
-	public void onProgress(int position);
+	public void onProgress(long position);
 }
